// src/resources/hotel/selectHotelUserInterfaces.ts
export enum selectHotelPage {

    city_result = "//h3[contains(text(),'Honolulu, HI, USA')]",
    filter_move = "#MainContentPlaceHolder_FilterFormLayout_NightlyRateTrackBar_MD",
    bar_width = "#MainContentPlaceHolder_FilterFormLayout_NightlyRateTrackBar_T",
    init_price = "#NightyRateTrackBarLabel_L",
    btn_aply = "#MainContentPlaceHolder_FilterFormLayout_ApplyFilterButton_CD",
    filterRating = "#MainContentPlaceHolder_FilterFormLayout_CustomerRatingTrackBar_MD",
    initRating = "#CustomerRatingTrackbarLabel_R",
    priceHotel = ".price",
    selecHotel = ".hotel-right-block"
   
}